Sets up trace exec
