+++ pbt/pyproject.toml
@@ -25,6 +25,7 @@ isort = "^5.9.3"
 
 [tool.poetry.scripts]
 bar = "pbt.bar:bar"
+foo = "pbt.foo:foo"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
