--- x/ccv/pbt/trace/hc.json
@@ -1,184 +1,0 @@
-[
-  {
-    "actions": [
-      {
-        "kind": "ProviderSlash",
-        "val": 1,
-        "power": 2,
-        "height": 0,
-        "factor": 0.25
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "provider",
-          "consumer"
-        ],
-        "n": 4,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "provider",
-          "consumer"
-        ],
-        "n": 4,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "consumer"
-        ],
-        "n": 4,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "ConsumerSlash",
-        "val": 2,
-        "power": 3,
-        "height": 6,
-        "is_downtime": true
-      },
-      {
-        "kind": "Undelegate",
-        "val": 1,
-        "amt": 4
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "provider",
-          "consumer"
-        ],
-        "n": 7,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "Deliver",
-        "chain": "provider"
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "provider"
-        ],
-        "n": 4,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "consumer"
-        ],
-        "n": 7,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "ConsumerSlash",
-        "val": 2,
-        "power": 3,
-        "height": 19,
-        "is_downtime": true
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "consumer"
-        ],
-        "n": 1,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "Undelegate",
-        "val": 3,
-        "amt": 1
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "provider",
-          "consumer"
-        ],
-        "n": 7,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "ConsumerSlash",
-        "val": 2,
-        "power": 3,
-        "height": 0,
-        "is_downtime": true
-      },
-      {
-        "kind": "Deliver",
-        "chain": "consumer"
-      },
-      {
-        "kind": "Undelegate",
-        "val": 3,
-        "amt": 2
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "consumer"
-        ],
-        "n": 7,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "Delegate",
-        "val": 1,
-        "amt": 1
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "consumer"
-        ],
-        "n": 1,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "Delegate",
-        "val": 3,
-        "amt": 2
-      },
-      {
-        "kind": "Deliver",
-        "chain": "provider"
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "consumer"
-        ],
-        "n": 7,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "JumpNBlocks",
-        "chains": [
-          "provider",
-          "consumer"
-        ],
-        "n": 4,
-        "seconds_per_block": 1
-      },
-      {
-        "kind": "Delegate",
-        "val": 1,
-        "amt": 2
-      },
-      {
-        "kind": "ProviderSlash",
-        "val": 1,
-        "power": 6,
-        "height": 23,
-        "factor": 0.25
-      }
-    ]
-  }
-]
